@@ -33,157 +33,39 @@
 		</developer>
 	</developers>
 
-
 	<properties>
-		<!-- kernel versions -->
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
 		<!-- maven -->
-		<maven.compiler.target>11</maven.compiler.target>
-		<maven.compiler.source>11</maven.compiler.source>
-		<maven.compiler.version>3.7.0</maven.compiler.version>
+		<maven.compiler.target>21</maven.compiler.target>
+		<maven.compiler.source>21</maven.compiler.source>
+		<maven.compiler.version>3.8.0</maven.compiler.version>
 		<maven.jar.plugin.version>3.0.2</maven.jar.plugin.version>
 		<maven.war.plugin.version>3.1.0</maven.war.plugin.version>
-		<maven.surefire.plugin.version>2.9</maven.surefire.plugin.version>
-		<maven.jacoco.plugin.version>0.8.1</maven.jacoco.plugin.version>
-		<maven.sonar.plugin.version>3.2</maven.sonar.plugin.version>
-		<maven.javadoc.version>3.0.1</maven.javadoc.version>
-		<maven-shade-plugin.version>2.3</maven-shade-plugin.version>
-		<maven.javadoc.skip>false</maven.javadoc.skip>
-		<maven.sonar.plugin.version>3.7.0.1746</maven.sonar.plugin.version>
-		<!-- spring -->
-		<spring.boot.version>2.0.2.RELEASE</spring.boot.version>
-		<spring.data.jpa.version>2.0.7.RELEASE</spring.data.jpa.version>
-		<spring.security.test.version>5.0.5.RELEASE</spring.security.test.version>
-		<spring-cloud-config.version>2.0.0.RELEASE</spring-cloud-config.version>
+		<maven.javadoc.version>3.2.0</maven.javadoc.version>
+		<maven.gpg.plugin.version>1.5</maven.gpg.plugin.version>
+		<maven.shade.plugin.version>3.2.4</maven.shade.plugin.version>
+		<maven.antrun.plugin.version>3.0.0</maven.antrun.plugin.version>
+		<maven.source.plugin.version>2.2.1</maven.source.plugin.version>
 
-		<swagger.core.version>2.0.7</swagger.core.version>
-		<swagger.annotations.version>1.5.20</swagger.annotations.version>
-		<springfox.version>2.9.2</springfox.version>
+		<git.commit.id.plugin.version>3.0.1</git.commit.id.plugin.version>
+		<fileName>apitest-mimoto-1.2.1-java21-SNAPSHOT-jar-with-dependencies</fileName>
+	</properties>
 
-		<vertx.version>3.6.2</vertx.version>
-		<vertx.health.check.version>3.7.0</vertx.health.check.version>
+	<dependencies>
+		<dependency>
+			<groupId>io.mosip.testrig.apirig.apitest.commons</groupId>
+			<artifactId>apitest-commons</artifactId>
+			<version>1.2.1-java21-SNAPSHOT</version>
+		</dependency>
+	</dependencies>
 
-
-		<!-- javax -->
-		<javax.interceptor.version>1.2</javax.interceptor.version>
-		<javax.el-api.version>3.0.0</javax.el-api.version>
-		<javax.transaction.api.version>1.3</javax.transaction.api.version>
-		<javax.persistence.api.version>2.2</javax.persistence.api.version>
-		<javax.validation-api.version>2.0.1.Final</javax.validation-api.version>
-		<glassfish.javax.el>2.2.6</glassfish.javax.el>
-
-		<!-- data -->
-		<h2.version>1.4.197</h2.version>
-		<mysql.connector.version>5.1.46</mysql.connector.version>
-		<postgresql.version>42.2.2</postgresql.version>
-		<eclipselink.version>2.5.0</eclipselink.version>
-		<hibernate.validator.version>6.0.12.Final</hibernate.validator.version>
-
-		<!-- test -->
-		<mockito.version>1.10.19</mockito.version>
-		<powermock.version>1.7.4</powermock.version>
-		<powermock.beta.version>2.0.0-beta.5</powermock.beta.version>
-
-		<!-- utilities -->
-		<commons-math3>3.6.1</commons-math3>
-		<commons-lang3>3.7</commons-lang3>
-		<commons-io>2.6</commons-io>
-		<commons-codec>1.11</commons-codec>
-		<commons-collections4.version>4.3</commons-collections4.version>
-		<commons.beanutils.version>1.9.2</commons.beanutils.version>
-		<commons.configuration2.version>2.2</commons.configuration2.version>
-		<apache.httpcomponents.version>4.5.6</apache.httpcomponents.version>
-		<guava.version>19.0</guava.version>
-		<lombok.version>1.18.8</lombok.version>
-		<jsch.version>0.1.54</jsch.version>
-		<sshd.version>1.4.0</sshd.version>
-		<itextcore.version>7.1.0</itextcore.version>
-		<itexthtml2pdf.version>2.0.0</itexthtml2pdf.version>
-		<itext.version>5.5.13</itext.version>
-		<freemarker.version>2.3.23</freemarker.version>
-		<velocity.version>1.7</velocity.version>
-		<velocity.tool.version>2.0</velocity.tool.version>
-		<datamapper.orika>1.5.2</datamapper.orika>
-		<modelmapper.version>2.1.1</modelmapper.version>
-		<bouncycastle.version>1.60</bouncycastle.version>
-		<icu4j.version>63.1</icu4j.version>
-		<seventeam-otp.version>1.0.0</seventeam-otp.version>
-		<google.zxing.version>3.3.3</google.zxing.version>
-		<hadoop-client>2.8.1</hadoop-client>
-		<htrace-core4>4.1.0-incubating</htrace-core4>
-		<aws.javasdk.version>1.11.368</aws.javasdk.version>
-		<s3mock.version>0.2.4</s3mock.version>
-		<ignite.version>2.3.0</ignite.version>
-		<findbugs.version>3.0.1</findbugs.version>
-		<byte.buddy.version>1.9.12</byte.buddy.version>
-		<io.jsonwebtoken.jjwt.version>0.6.0</io.jsonwebtoken.jjwt.version>
-		<apache.directory.api.version>2.0.0.AM2</apache.directory.api.version>
-		
-		
-		
-
-
-
-		<!-- registration versions -->
-		<maven.jacoco.version>0.8.2</maven.jacoco.version>
-		<javafx.version>2.2</javafx.version>
-		<maven.sonar.version>3.2</maven.sonar.version>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<maven.compiler.version>3.7.0</maven.compiler.version>
-		<maven.javadoc.version>3.0.1</maven.javadoc.version>
-
-		<!-- Test & Logging -->
-		<logback.version>1.1.6</logback.version>
-		<slf4j.version>1.7.19</slf4j.version>
-		<mockito.version>1.10.19</mockito.version>
-
-		<!-- Orika Mapper -->
-		<orika.version>1.4.6</orika.version>
-
-		<!-- Spring version -->
-		<spring.version>5.0.6.RELEASE</spring.version>
-
-		<!-- Hibernate version -->
-		<hibernate.version>5.4.2.Final</hibernate.version>
-
-		<!-- Apache Commons DPC2 version -->
-		<apache.commons.dpce.version>2.1.1</apache.commons.dpce.version>
-
-		<!-- Derby Version -->
-		<apache.derby.version>10.13.1.1</apache.derby.version>
-
-		<!-- sourceafis Version -->
-		<sourceafis.version>3.4.0</sourceafis.version>
-
-		<!-- openimaj version -->
-		<org.openimaj.version>1.3.5</org.openimaj.version>
-
-		<!-- webcam-capture version -->
-		<webcam-capture.version>0.3.12</webcam-capture.version>
-
-		<!-- TSS Version -->
-		<tss.version>0.3.0</tss.version>
-
-		<!-- automationtests version -->
-		<wink-json4j-provider.version>1.1.2-incubating</wink-json4j-provider.version>
-		<kernel.version>1.2.0.1-B1</kernel.version>
-		<zjsonpatch.version>0.4.7</zjsonpatch.version>
-		<extentreports.aventstack.version>3.0.0</extentreports.aventstack.version>
-		<extentreports.relevantcodes.version>2.41.2</extentreports.relevantcodes.version>
-		<json-path.version>2.4.0</json-path.version>
-		<maven.model.version>3.3.9</maven.model.version>
-		<rest.assured.version>3.0.7</rest.assured.version>
-		<testng.version>6.11</testng.version>
-		<zt.zip.version>1.13</zt.zip.version>
-		<fileName>apitest-mimoto-1.2.0.1-SNAPSHOT-jar-with-dependencies</fileName>
-		
-		</properties>
 	<build>
 		<pluginManagement>
 			<plugins>
 			</plugins>
 		</pluginManagement>
+
 		<plugins>
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
