@@ -1,11 +1,17 @@
 package io.mosip.mimoto.dto.mimoto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 @Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
 public class VCCredentialResponseProof {
     @NotBlank
     private String type;
