@@ -5,6 +5,7 @@ import com.google.gson.annotations.SerializedName;
 import lombok.Data;
 
 import java.util.List;
+import java.util.Map;
 
 @Data
 public class CredentialIssuerWellKnownResponse {
@@ -12,11 +13,15 @@ public class CredentialIssuerWellKnownResponse {
     @JsonProperty("credential_issuer")
     private String credentialIssuer;
 
+    @SerializedName("authorization_servers")
+    @JsonProperty("authorization_servers")
+    private List<String> authorizationServers;
+
     @SerializedName("credential_endpoint")
     @JsonProperty("credential_endpoint")
     private String credentialEndPoint;
 
-    @SerializedName("credentials_supported")
-    @JsonProperty("credentials_supported")
-    private List<CredentialsSupportedResponse> credentialsSupported;
+    @SerializedName("credential_configurations_supported")
+    @JsonProperty("credential_configurations_supported")
+    private Map<String,CredentialsSupportedResponse> credentialConfigurationsSupported;
 }
