@@ -5,9 +5,7 @@ import io.mosip.mimoto.dto.IssuersDTO;
 import io.mosip.mimoto.dto.mimoto.*;
 import io.mosip.mimoto.exception.ApiNotAccessibleException;
 import io.mosip.mimoto.exception.InvalidIssuerIdException;
-import org.apache.http.auth.InvalidCredentialsException;
 
-import java.io.ByteArrayInputStream;
 import java.io.IOException;
 
 public interface IssuersService {
@@ -16,4 +14,6 @@ public interface IssuersService {
     IssuersDTO getAllIssuersWithAllFields() throws ApiNotAccessibleException, IOException;
 
     IssuerDTO getIssuerConfig(String issuerId) throws ApiNotAccessibleException, IOException, InvalidIssuerIdException;
+
+    CredentialIssuerWellKnownResponse getIssuerWellknown(String issuerId) throws ApiNotAccessibleException, IOException;
 }
