@@ -1,5 +1,6 @@
 package io.mosip.mimoto.service.impl;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.zxing.BarcodeFormat;
 import com.google.zxing.WriterException;
@@ -15,8 +16,11 @@ import io.mosip.mimoto.dto.IssuerDTO;
 import io.mosip.mimoto.dto.IssuersDTO;
 import io.mosip.mimoto.dto.idp.TokenResponseDTO;
 import io.mosip.mimoto.dto.mimoto.*;
+import io.mosip.mimoto.dto.openid.presentation.PresentationDefinitionDTO;
 import io.mosip.mimoto.exception.ApiNotAccessibleException;
 import io.mosip.mimoto.exception.IdpException;
+import io.mosip.mimoto.exception.InvalidCredentialResourceException;
+import io.mosip.mimoto.exception.ErrorConstants;
 import io.mosip.mimoto.service.CredentialService;
 import io.mosip.mimoto.service.IdpService;
 import io.mosip.mimoto.service.IssuersService;
@@ -26,10 +30,11 @@ import io.mosip.mimoto.util.RestApiClient;
 import io.mosip.mimoto.util.Utilities;
 import io.mosip.pixelpass.PixelPass;
 import org.apache.commons.lang.StringUtils;
-import org.apache.http.auth.InvalidCredentialsException;
 import org.apache.velocity.VelocityContext;
 import org.apache.velocity.app.Velocity;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
@@ -42,10 +47,10 @@ import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.StringWriter;
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 import java.util.*;
 import java.util.stream.Collectors;
-
-import static io.mosip.mimoto.exception.PlatformErrorMessages.INVALID_CREDENTIAL_TYPE_EXCEPTION;
 
 @Service
 public class CredentialServiceImpl implements CredentialService {
@@ -65,7 +70,19 @@ public class CredentialServiceImpl implements CredentialService {
     IssuersService issuerService;
 
     @Autowired
+    DataShareServiceImpl dataShareService;
+
+    @Autowired
+    ObjectMapper objectMapper;
+
+    @Autowired
     IdpService idpService;
+
+    @Value("${mosip.inji.web.authorize.url}")
+    String injiWebAuthorizeUrl;
+
+    @Autowired
+    PresentationServiceImpl presentationService;
 
     @Override
     public TokenResponseDTO getTokenResponse(Map<String, String> params, String issuerId) throws ApiNotAccessibleException, IOException {
@@ -82,14 +99,15 @@ public class CredentialServiceImpl implements CredentialService {
     @Override
     public ByteArrayInputStream downloadCredentialAsPDF(String issuerId, String credentialType, TokenResponseDTO response) throws Exception {
         IssuerDTO issuerConfig = issuerService.getIssuerConfig(issuerId);
-        CredentialIssuerWellKnownResponse credentialIssuerWellKnownResponse = getCredentialIssuerWellknown(issuerId, credentialType);
-        CredentialsSupportedResponse credentialsSupportedResponse = getCredentialSupported(credentialIssuerWellKnownResponse, credentialType);
-        VCCredentialRequest vcCredentialRequest = generateVCCredentialRequest(issuerConfig, credentialsSupportedResponse, response.getAccess_token());
-        VCCredentialResponse vcCredentialResponse = downloadCredential(credentialIssuerWellKnownResponse.getCredentialEndPoint(), vcCredentialRequest, response.getAccess_token());
-        return generatePdfForVerifiableCredentials(vcCredentialResponse, issuerConfig, credentialsSupportedResponse, credentialIssuerWellKnownResponse.getCredentialEndPoint());
+        CredentialIssuerWellKnownResponseDraft11 credentialIssuerWellKnownResponseDraft11 = getCredentialIssuerWellknown(issuerId, credentialType);
+        CredentialsSupportedResponseDraft11 credentialsSupportedResponseDraft11 = getCredentialSupported(credentialIssuerWellKnownResponseDraft11, credentialType);
+        VCCredentialRequest vcCredentialRequest = generateVCCredentialRequest(issuerConfig, credentialsSupportedResponseDraft11, response.getAccess_token());
+        VCCredentialResponse vcCredentialResponse = downloadCredential(credentialIssuerWellKnownResponseDraft11.getCredentialEndPoint(), vcCredentialRequest, response.getAccess_token());
+        String dataShareUrl = dataShareService.storeDataInDataShare(objectMapper.writeValueAsString(vcCredentialResponse));
+        return generatePdfForVerifiableCredentials(vcCredentialResponse, issuerConfig, credentialsSupportedResponseDraft11, dataShareUrl);
     }
 
-    public VCCredentialResponse downloadCredential(String credentialEndpoint, VCCredentialRequest vcCredentialRequest, String accessToken) throws ApiNotAccessibleException, IOException, InvalidCredentialsException {
+    public VCCredentialResponse downloadCredential(String credentialEndpoint, VCCredentialRequest vcCredentialRequest, String accessToken) throws InvalidCredentialResourceException {
         VCCredentialResponse vcCredentialResponse = restApiClient.postApi(credentialEndpoint, MediaType.APPLICATION_JSON,
                 vcCredentialRequest, VCCredentialResponse.class, accessToken);
         logger.debug("VC Credential Response is -> " + vcCredentialResponse);
@@ -97,31 +115,37 @@ public class CredentialServiceImpl implements CredentialService {
         return vcCredentialResponse;
     }
 
-    public VCCredentialRequest generateVCCredentialRequest(IssuerDTO issuerDTO, CredentialsSupportedResponse credentialsSupportedResponse, String accessToken) throws Exception {
+    public VCCredentialRequest generateVCCredentialRequest(IssuerDTO issuerDTO, CredentialsSupportedResponseDraft11 credentialsSupportedResponseDraft11, String accessToken) throws Exception {
         String jwt = joseUtil.generateJwt(issuerDTO.getCredential_audience(), issuerDTO.getClient_id(), accessToken);
         return VCCredentialRequest.builder()
-                .format(credentialsSupportedResponse.getFormat())
+                .format(credentialsSupportedResponseDraft11.getFormat())
                 .proof(VCCredentialRequestProof.builder()
-                        .proofType(credentialsSupportedResponse.getProofTypesSupported().get(0))
+                        .proofType(credentialsSupportedResponseDraft11.getProofTypesSupported().get(0))
                         .jwt(jwt)
                         .build())
                 .credentialDefinition(VCCredentialDefinition.builder()
-                        .type(credentialsSupportedResponse.getCredentialDefinition().getType())
+                        .type(credentialsSupportedResponseDraft11.getCredentialDefinition().getType())
                         .context(List.of("https://www.w3.org/2018/credentials/v1"))
                         .build())
                 .build();
     }
 
-    public ByteArrayInputStream generatePdfForVerifiableCredentials(VCCredentialResponse vcCredentialResponse, IssuerDTO issuerDTO, CredentialsSupportedResponse credentialsSupportedResponse, String credentialEndPoint) throws Exception {
+    public ByteArrayInputStream generatePdfForVerifiableCredentials(VCCredentialResponse vcCredentialResponse, IssuerDTO issuerDTO, CredentialsSupportedResponseDraft11 credentialsSupportedResponseDraft11, String dataShareUrl) throws Exception {
+        LinkedHashMap<String, Object> displayProperties = loadDisplayPropertiesFromWellknown(vcCredentialResponse, credentialsSupportedResponseDraft11);
+        Map<String, Object> data = getPdfResourceFromVcProperties(displayProperties, credentialsSupportedResponseDraft11,  vcCredentialResponse, issuerDTO, dataShareUrl);
+        return renderVCInCredentialTemplate(data);
+    }
 
+    @NotNull
+    private static LinkedHashMap<String, Object> loadDisplayPropertiesFromWellknown(VCCredentialResponse vcCredentialResponse, CredentialsSupportedResponseDraft11 credentialsSupportedResponseDraft11) {
         LinkedHashMap<String,Object> displayProperties = new LinkedHashMap<>();
         Map<String, Object> credentialProperties = vcCredentialResponse.getCredential().getCredentialSubject();
 
         LinkedHashMap<String, String> vcPropertiesFromWellKnown = new LinkedHashMap<>();
-        Map<String, CredentialDisplayResponseDto> credentialSubject = credentialsSupportedResponse.getCredentialDefinition().getCredentialSubject();
+        Map<String, CredentialDisplayResponseDto> credentialSubject = credentialsSupportedResponseDraft11.getCredentialDefinition().getCredentialSubject();
         credentialSubject.keySet().forEach(VCProperty -> vcPropertiesFromWellKnown.put(VCProperty, credentialSubject.get(VCProperty).getDisplay().get(0).getName()));
 
-        Set<String> orderProperty = credentialsSupportedResponse.getOrder();
+        Set<String> orderProperty = credentialsSupportedResponseDraft11.getOrder();
 
         Set<String> fieldProperties = orderProperty == null ? vcPropertiesFromWellKnown.keySet() : orderProperty;
         fieldProperties.forEach(vcProperty -> {
@@ -129,17 +153,16 @@ public class CredentialServiceImpl implements CredentialService {
                 displayProperties.put(vcPropertiesFromWellKnown.get(vcProperty), credentialProperties.get(vcProperty));
             }
         });
-        return getPdfResourceFromVcProperties(displayProperties, credentialsSupportedResponse,  vcCredentialResponse,
-                issuerDTO.getDisplay().stream().map(d -> d.getLogo().getUrl()).findFirst().orElse(""));
+        return displayProperties;
     }
 
 
-    private ByteArrayInputStream getPdfResourceFromVcProperties(LinkedHashMap<String, Object> displayProperties, CredentialsSupportedResponse credentialsSupportedResponse, VCCredentialResponse  vcCredentialResponse, String issuerLogoUrl) throws IOException, WriterException {
+    private Map<String, Object> getPdfResourceFromVcProperties(LinkedHashMap<String, Object> displayProperties, CredentialsSupportedResponseDraft11 credentialsSupportedResponseDraft11, VCCredentialResponse  vcCredentialResponse, IssuerDTO issuerDTO, String dataShareUrl) throws IOException, WriterException {
         Map<String, Object> data = new HashMap<>();
         LinkedHashMap<String, Object> rowProperties = new LinkedHashMap<>();
-        String backgroundColor = credentialsSupportedResponse.getDisplay().get(0).getBackgroundColor();
-        String textColor = credentialsSupportedResponse.getDisplay().get(0).getTextColor();
-        String credentialSupportedType = credentialsSupportedResponse.getDisplay().get(0).getName();
+        String backgroundColor = credentialsSupportedResponseDraft11.getDisplay().get(0).getBackgroundColor();
+        String textColor = credentialsSupportedResponseDraft11.getDisplay().get(0).getTextColor();
+        String credentialSupportedType = credentialsSupportedResponseDraft11.getDisplay().get(0).getName();
         String face = vcCredentialResponse.getCredential().getCredentialSubject().get("face") != null ? (String) vcCredentialResponse.getCredential().getCredentialSubject().get("face") : null;
 
         displayProperties.entrySet().stream()
@@ -159,27 +182,21 @@ public class CredentialServiceImpl implements CredentialService {
                     }
                 });
 
-        if(!credentialsSupportedResponse.getId().equals("MOSIPVerifiableCredential")) {
-            PixelPass pixelPass = new PixelPass();
-            ObjectMapper objectMapper = new ObjectMapper();
-            logger.info("Credential That is converted to PDF" + objectMapper.writeValueAsString(vcCredentialResponse.getCredential()));
-            String qrData = pixelPass.generateQRData(objectMapper.writeValueAsString(vcCredentialResponse.getCredential()), "");
-            logger.info("QR Data => " + qrData);
-            QRCodeWriter qrCodeWriter = new QRCodeWriter();
-            BitMatrix bitMatrix = qrCodeWriter.encode(qrData, BarcodeFormat.QR_CODE, 650, 650);
-            BufferedImage qrImage = MatrixToImageWriter.toBufferedImage(bitMatrix);
-            String base64Image = encodeToString(qrImage, "png");
-            data.put("qrCodeImage", base64Image);
-        }
-
-
-        data.put("logoUrl", issuerLogoUrl);
+        String qrCodeImage = !"false".equals(issuerDTO.getOvp_qr_enabled()) ?
+                constructQRCodeWithAuthorizeRequest(vcCredentialResponse, dataShareUrl) :
+                constructQRCodeWithVCData(credentialsSupportedResponseDraft11, vcCredentialResponse) ;
+        data.put("qrCodeImage", qrCodeImage);
+        data.put("logoUrl", issuerDTO.getDisplay().stream().map(d -> d.getLogo().getUrl()).findFirst().orElse(""));
         data.put("rowProperties", rowProperties);
         data.put("textColor", textColor);
         data.put("backgroundColor", backgroundColor);
         data.put("titleName", credentialSupportedType);
         data.put("face", face);
+        return data;
+    }
 
+    @NotNull
+    private ByteArrayInputStream renderVCInCredentialTemplate(Map<String, Object> data) throws IOException {
         String  credentialTemplate = utilities.getCredentialSupportedTemplateString();
 
         Properties props = new Properties();
@@ -203,6 +220,31 @@ public class CredentialServiceImpl implements CredentialService {
         HtmlConverter.convertToPdf(mergedHtml, pdfwriter, converterProperties);
         return new ByteArrayInputStream(outputStream.toByteArray());
     }
+
+    private String constructQRCode(String qrData) throws WriterException {
+        QRCodeWriter qrCodeWriter = new QRCodeWriter();
+        BitMatrix bitMatrix = qrCodeWriter.encode(qrData, BarcodeFormat.QR_CODE, 650, 650);
+        BufferedImage qrImage = MatrixToImageWriter.toBufferedImage(bitMatrix);
+        return encodeToString(qrImage, "png");
+    }
+
+    private String constructQRCodeWithVCData(CredentialsSupportedResponseDraft11 credentialsSupportedResponseDraft11, VCCredentialResponse vcCredentialResponse) throws JsonProcessingException, WriterException {
+        if(!credentialsSupportedResponseDraft11.getId().equals("MOSIPVerifiableCredential")) {
+            PixelPass pixelPass = new PixelPass();
+            ObjectMapper objectMapper = new ObjectMapper();
+            String qrData = pixelPass.generateQRData(objectMapper.writeValueAsString(vcCredentialResponse.getCredential()), "");
+            return constructQRCode(qrData);
+        }
+        return "";
+    }
+    private String constructQRCodeWithAuthorizeRequest(VCCredentialResponse vcCredentialResponse, String dataShareUrl) throws WriterException, JsonProcessingException {
+        PresentationDefinitionDTO presentationDefinitionDTO = presentationService.constructPresentationDefinition(vcCredentialResponse);
+        ObjectMapper objectMapper = new ObjectMapper();
+        String presentationString = objectMapper.writeValueAsString(presentationDefinitionDTO);
+        String qrData = String.format(injiWebAuthorizeUrl, URLEncoder.encode(dataShareUrl, StandardCharsets.UTF_8), URLEncoder.encode(presentationString, StandardCharsets.UTF_8));
+        return constructQRCode(qrData);
+    }
+
     public static String encodeToString(BufferedImage image, String type) {
         String imageString = null;
         ByteArrayOutputStream bos = new ByteArrayOutputStream();
@@ -231,11 +273,11 @@ public class CredentialServiceImpl implements CredentialService {
         if (issuerConfigResp.isPresent()) {
             IssuerDTO issuerDto = issuerConfigResp.get();
 
-            CredentialIssuerWellKnownResponse response = restApiClient.getApi(issuerDto.getWellKnownEndpoint(), CredentialIssuerWellKnownResponse.class);
+            CredentialIssuerWellKnownResponseDraft11 response = restApiClient.getApi(issuerDto.getWellKnownEndpoint(), CredentialIssuerWellKnownResponseDraft11.class);
             if (response == null) {
                 throw new ApiNotAccessibleException();
             }
-            List<CredentialsSupportedResponse> issuerCredentialsSupported = response.getCredentialsSupported();
+            List<CredentialsSupportedResponseDraft11> issuerCredentialsSupported = response.getCredentialsSupported();
             credentialTypesWithAuthorizationEndpoint.setAuthorizationEndPoint(issuerDto.getAuthorization_endpoint());
             credentialTypesWithAuthorizationEndpoint.setSupportedCredentials(issuerCredentialsSupported);
 
@@ -250,29 +292,29 @@ public class CredentialServiceImpl implements CredentialService {
         }
         return credentialTypesWithAuthorizationEndpoint;
     }
-    public CredentialIssuerWellKnownResponse getCredentialIssuerWellknown(String issuerId, String search) throws ApiNotAccessibleException, IOException {
-        CredentialIssuerWellKnownResponse credentialIssuerWellKnownResponse = new CredentialIssuerWellKnownResponse();
+    public CredentialIssuerWellKnownResponseDraft11 getCredentialIssuerWellknown(String issuerId, String search) throws ApiNotAccessibleException, IOException {
+        CredentialIssuerWellKnownResponseDraft11 credentialIssuerWellKnownResponseDraft11 = new CredentialIssuerWellKnownResponseDraft11();
         IssuersDTO issuersDto = issuerService.getAllIssuersWithAllFields();
         Optional<IssuerDTO> issuerConfigResp = issuersDto.getIssuers().stream()
                 .filter(issuer -> issuer.getCredential_issuer().equals(issuerId))
                 .findFirst();
         if (issuerConfigResp.isPresent()) {
             IssuerDTO issuerDto = issuerConfigResp.get();
-            credentialIssuerWellKnownResponse = restApiClient.getApi(issuerDto.getWellKnownEndpoint(), CredentialIssuerWellKnownResponse.class);
-            if (credentialIssuerWellKnownResponse == null) {
+            credentialIssuerWellKnownResponseDraft11 = restApiClient.getApi(issuerDto.getWellKnownEndpoint(), CredentialIssuerWellKnownResponseDraft11.class);
+            if (credentialIssuerWellKnownResponseDraft11 == null) {
                 throw new ApiNotAccessibleException();
             }
         }
-        return credentialIssuerWellKnownResponse;
+        return credentialIssuerWellKnownResponseDraft11;
     }
 
-    public CredentialsSupportedResponse getCredentialSupported(CredentialIssuerWellKnownResponse credentialIssuerWellKnownResponse, String credentialType) throws ApiNotAccessibleException, IOException, InvalidCredentialsException {
-        Optional<CredentialsSupportedResponse> credentialsSupportedResponse = credentialIssuerWellKnownResponse.getCredentialsSupported().stream()
+    public CredentialsSupportedResponseDraft11 getCredentialSupported(CredentialIssuerWellKnownResponseDraft11 credentialIssuerWellKnownResponseDraft11, String credentialType) throws InvalidCredentialResourceException {
+        Optional<CredentialsSupportedResponseDraft11> credentialsSupportedResponse = credentialIssuerWellKnownResponseDraft11.getCredentialsSupported().stream()
                 .filter(credentialsSupported -> credentialsSupported.getId().equals(credentialType))
                 .findFirst();
         if (credentialsSupportedResponse.isEmpty()){
             logger.error("Invalid credential Type passed - {}", credentialType);
-            throw new InvalidCredentialsException(INVALID_CREDENTIAL_TYPE_EXCEPTION.getMessage());
+            throw new InvalidCredentialResourceException(ErrorConstants.RESOURCE_NOT_FOUND.getErrorMessage());
         }
         return credentialsSupportedResponse.get();
     }
