@@ -6,6 +6,7 @@ import com.google.gson.GsonBuilder;
 import io.mosip.kernel.core.logger.spi.Logger;
 import io.mosip.mimoto.dto.IssuerDTO;
 import io.mosip.mimoto.dto.IssuersDTO;
+import io.mosip.mimoto.dto.mimoto.CredentialIssuerWellKnownResponse;
 import io.mosip.mimoto.exception.ApiNotAccessibleException;
 import io.mosip.mimoto.exception.InvalidIssuerIdException;
 import io.mosip.mimoto.service.IssuersService;
@@ -15,8 +16,10 @@ import io.mosip.mimoto.util.Utilities;
 import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import java.io.*;
-import java.util.*;
+
+import java.io.IOException;
+import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 
@@ -26,6 +29,10 @@ public class IssuersServiceImpl implements IssuersService {
 
     @Autowired
     private Utilities utilities;
+
+    @Autowired
+    private RestApiClient restApiClient;
+
 
     @Override
     public IssuersDTO getAllIssuers(String search) throws ApiNotAccessibleException, IOException {
@@ -67,7 +74,6 @@ public class IssuersServiceImpl implements IssuersService {
     }
 
 
-
     @Override
     public IssuerDTO getIssuerConfig(String issuerId) throws ApiNotAccessibleException, IOException {
         IssuerDTO issuerDTO = null;
@@ -85,4 +91,14 @@ public class IssuersServiceImpl implements IssuersService {
             throw new InvalidIssuerIdException();
         return issuerDTO;
     }
+
+    @Override
+    public CredentialIssuerWellKnownResponse getIssuerWellknown(String issuerId) throws ApiNotAccessibleException, IOException {
+        IssuerDTO issuerConfig = getIssuerConfig(issuerId);
+        String wellknown = issuerConfig.getWellKnownEndpoint();
+        String wellknownJson = restApiClient.getApi(wellknown, String.class);
+        return new Gson().fromJson(wellknownJson, CredentialIssuerWellKnownResponse.class);
+    }
 }
+
+
